create separate crate, address review comments
