+++ consensus/Cargo.toml
@@ -1,0 +1,53 @@
+[package]
+name = "solana-consensus"
+description = "Blockchain, Rebuilt for Scale"
+documentation = "https://docs.rs/solana-consensus"
+version = { workspace = true }
+authors = { workspace = true }
+repository = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+edition = { workspace = true }
+publish = false
+
+[badges]
+codecov = { repository = "solana-labs/solana", branch = "master", service = "github" }
+
+[dependencies]
+bincode = { workspace = true }
+chrono = { workspace = true, features = ["default", "serde"] }
+crossbeam-channel = { workspace = true }
+eager = { workspace = true }
+etcd-client = { workspace = true, features = ["tls"] }
+itertools = { workspace = true }
+log = { workspace = true }
+serde = { workspace = true }
+serde_derive = { workspace = true }
+solana-frozen-abi = { workspace = true }
+solana-frozen-abi-macro = { workspace = true }
+solana-gossip = { workspace = true }
+solana-ledger = { workspace = true }
+solana-measure = { workspace = true }
+solana-metrics = { workspace = true }
+solana-program-runtime = { workspace = true }
+solana-runtime = { workspace = true }
+solana-sdk = { workspace = true }
+solana-streamer = { workspace = true }
+solana-vote = { workspace = true }
+solana-vote-program = { workspace = true }
+tempfile = { workspace = true }
+thiserror = { workspace = true }
+tokio = { workspace = true, features = ["full"] }
+trees = { workspace = true }
+
+[dev-dependencies]
+assert_matches = { workspace = true }
+solana-consensus = { path = "." }
+solana-core = { workspace = true }
+solana-logger = { workspace = true }
+
+[build-dependencies]
+rustc_version = { workspace = true }
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
