+++ consensus/src/cluster_slots_service.rs
@@ -1,5 +1,5 @@
 use {
-    crate::consensus::cluster_slots::ClusterSlots,
+    crate::cluster_slots::ClusterSlots,
     crossbeam_channel::{Receiver, RecvTimeoutError, Sender},
     solana_gossip::cluster_info::ClusterInfo,
     solana_ledger::blockstore::Blockstore,
