+++ consensus/src/tower_storage.rs
@@ -1,7 +1,8 @@
 use {
-    crate::consensus::{
-        tower1_14_11::Tower1_14_11, tower1_7_14::SavedTower1_7_14, Result, Tower, TowerError,
-        TowerVersions,
+    crate::{
+        consensus::{Result, Tower, TowerError, TowerVersions},
+        tower1_14_11::Tower1_14_11,
+        tower1_7_14::SavedTower1_7_14,
     },
     solana_sdk::{
         pubkey::Pubkey,
@@ -370,9 +371,9 @@ impl TowerStorage for EtcdTowerStorage {
 pub mod test {
     use {
         super::*,
-        crate::consensus::{
+        crate::{
+            consensus::Tower,
             tower1_7_14::{SavedTower1_7_14, Tower1_7_14},
-            Tower,
         },
         solana_sdk::{hash::Hash, signature::Keypair},
         solana_vote_program::vote_state::{
