+++ core/src/repair/cluster_slot_state_verifier.rs
@@ -1,11 +1,9 @@
 use {
-    crate::{
-        consensus::{
-            fork_choice::ForkChoice, heaviest_subtree_fork_choice::HeaviestSubtreeForkChoice,
-        },
-        repair::ancestor_hashes_service::{
-            AncestorHashesReplayUpdate, AncestorHashesReplayUpdateSender,
-        },
+    crate::repair::ancestor_hashes_service::{
+        AncestorHashesReplayUpdate, AncestorHashesReplayUpdateSender,
+    },
+    solana_consensus::{
+        fork_choice::ForkChoice, heaviest_subtree_fork_choice::HeaviestSubtreeForkChoice,
     },
     solana_ledger::blockstore::Blockstore,
     solana_sdk::{clock::Slot, hash::Hash},
@@ -946,8 +944,9 @@ pub(crate) fn check_slot_agrees_with_cluster(
 mod test {
     use {
         super::*,
-        crate::{consensus::progress_map::ProgressMap, replay_stage::tests::setup_forks_from_tree},
+        crate::replay_stage::tests::setup_forks_from_tree,
         crossbeam_channel::unbounded,
+        solana_consensus::progress_map::ProgressMap,
         solana_runtime::bank_forks::BankForks,
         std::{
             collections::{HashMap, HashSet},
