+++ core/src/repair/repair_weight.rs
@@ -1,13 +1,13 @@
 use {
-    crate::{
-        consensus::{heaviest_subtree_fork_choice::HeaviestSubtreeForkChoice, tree_diff::TreeDiff},
-        repair::{
-            repair_generic_traversal::{get_closest_completion, get_unknown_last_index},
-            repair_service::{BestRepairsStats, RepairTiming},
-            repair_weighted_traversal,
-            serve_repair::ShredRepairType,
-        },
-        replay_stage::DUPLICATE_THRESHOLD,
+    crate::repair::{
+        repair_generic_traversal::{get_closest_completion, get_unknown_last_index},
+        repair_service::{BestRepairsStats, RepairTiming},
+        repair_weighted_traversal,
+        serve_repair::ShredRepairType,
+    },
+    solana_consensus::{
+        consensus::DUPLICATE_THRESHOLD, heaviest_subtree_fork_choice::HeaviestSubtreeForkChoice,
+        tree_diff::TreeDiff,
     },
     solana_ledger::{
         ancestor_iterator::AncestorIterator, blockstore::Blockstore, blockstore_meta::SlotMeta,
