+++ core/src/unfrozen_gossip_verified_vote_hashes.rs
@@ -1,5 +1,5 @@
 use {
-    crate::consensus::latest_validator_votes_for_frozen_banks::LatestValidatorVotesForFrozenBanks,
+    solana_consensus::latest_validator_votes_for_frozen_banks::LatestValidatorVotesForFrozenBanks,
     solana_sdk::{clock::Slot, hash::Hash, pubkey::Pubkey},
     std::collections::{BTreeMap, HashMap},
 };
