+++ core/src/vote_simulator.rs
@@ -1,22 +1,21 @@
 #![cfg(feature = "dev-context-only-utils")]
 use {
     crate::{
-        cluster_info_vote_listener::VoteTracker,
-        consensus::{
-            cluster_slots::ClusterSlots,
-            fork_choice::{HeaviestForkFailures, SelectVoteAndResetForkResult},
-            heaviest_subtree_fork_choice::HeaviestSubtreeForkChoice,
-            latest_validator_votes_for_frozen_banks::LatestValidatorVotesForFrozenBanks,
-            progress_map::{ForkProgress, ProgressMap},
-            Tower,
-        },
         repair::cluster_slot_state_verifier::{
             DuplicateConfirmedSlots, DuplicateSlotsTracker, EpochSlotsFrozenSlots,
         },
-        replay_stage::ReplayStage,
+        replay_stage::{HeaviestForkFailures, ReplayStage, SelectVoteAndResetForkResult},
         unfrozen_gossip_verified_vote_hashes::UnfrozenGossipVerifiedVoteHashes,
     },
     crossbeam_channel::unbounded,
+    solana_consensus::{
+        cluster_slots::ClusterSlots,
+        consensus::Tower,
+        heaviest_subtree_fork_choice::HeaviestSubtreeForkChoice,
+        latest_validator_votes_for_frozen_banks::LatestValidatorVotesForFrozenBanks,
+        progress_map::{ForkProgress, ProgressMap},
+        vote_stake_tracker::VoteTracker,
+    },
     solana_runtime::{
         accounts_background_service::AbsRequestSender,
         bank::Bank,
@@ -376,880 +375,4 @@ pub fn initialize_state(
     let heaviest_subtree_fork_choice =
         HeaviestSubtreeForkChoice::new_from_bank_forks(bank_forks.clone());
     (bank_forks, progress, heaviest_subtree_fork_choice)
-}
-
-#[cfg(test)]
-mod test {
-    use {
-        super::*,
-        crate::{
-            consensus::{fork_choice::ForkChoice, tree_diff::TreeDiff, SwitchForkDecision},
-            vote_simulator::VoteSimulator,
-        },
-        itertools::Itertools,
-        solana_sdk::{hash::Hash, slot_history::SlotHistory},
-        trees::tr,
-    };
-
-    #[test]
-    fn test_new_from_frozen_banks() {
-        /*
-            Build fork structure:
-                 slot 0
-                   |
-                 slot 1
-                 /    \
-            slot 2    |
-               |    slot 3
-            slot 4
-        */
-        let forks = tr(0) / (tr(1) / (tr(2) / (tr(4))) / (tr(3)));
-        let mut vote_simulator = VoteSimulator::new(1);
-        vote_simulator.fill_bank_forks(forks, &HashMap::new(), true);
-        let bank_forks = vote_simulator.bank_forks;
-        let mut frozen_banks: Vec<_> = bank_forks
-            .read()
-            .unwrap()
-            .frozen_banks()
-            .values()
-            .cloned()
-            .collect();
-        frozen_banks.sort_by_key(|bank| bank.slot());
-
-        let root_bank = bank_forks.read().unwrap().root_bank();
-        let root = root_bank.slot();
-        let root_hash = root_bank.hash();
-        let heaviest_subtree_fork_choice =
-            HeaviestSubtreeForkChoice::new_from_frozen_banks((root, root_hash), &frozen_banks);
-
-        let bank0_hash = bank_forks.read().unwrap().get(0).unwrap().hash();
-        assert!(heaviest_subtree_fork_choice
-            .parent(&(0, bank0_hash))
-            .is_none());
-
-        let bank1_hash = bank_forks.read().unwrap().get(1).unwrap().hash();
-        assert_eq!(
-            (&heaviest_subtree_fork_choice)
-                .children(&(0, bank0_hash))
-                .unwrap()
-                .collect_vec(),
-            &[&(1, bank1_hash)]
-        );
-
-        assert_eq!(
-            heaviest_subtree_fork_choice.parent(&(1, bank1_hash)),
-            Some((0, bank0_hash))
-        );
-        let bank2_hash = bank_forks.read().unwrap().get(2).unwrap().hash();
-        let bank3_hash = bank_forks.read().unwrap().get(3).unwrap().hash();
-        assert_eq!(
-            (&heaviest_subtree_fork_choice)
-                .children(&(1, bank1_hash))
-                .unwrap()
-                .collect_vec(),
-            &[&(2, bank2_hash), &(3, bank3_hash)]
-        );
-        assert_eq!(
-            heaviest_subtree_fork_choice.parent(&(2, bank2_hash)),
-            Some((1, bank1_hash))
-        );
-        let bank4_hash = bank_forks.read().unwrap().get(4).unwrap().hash();
-        assert_eq!(
-            (&heaviest_subtree_fork_choice)
-                .children(&(2, bank2_hash))
-                .unwrap()
-                .collect_vec(),
-            &[&(4, bank4_hash)]
-        );
-        // Check parent and children of invalid hash don't exist
-        let invalid_hash = Hash::new_unique();
-        assert!((&heaviest_subtree_fork_choice)
-            .children(&(2, invalid_hash))
-            .is_none());
-        assert!(heaviest_subtree_fork_choice
-            .parent(&(2, invalid_hash))
-            .is_none());
-
-        assert_eq!(
-            heaviest_subtree_fork_choice.parent(&(3, bank3_hash)),
-            Some((1, bank1_hash))
-        );
-        assert!((&heaviest_subtree_fork_choice)
-            .children(&(3, bank3_hash))
-            .unwrap()
-            .collect_vec()
-            .is_empty());
-        assert_eq!(
-            heaviest_subtree_fork_choice.parent(&(4, bank4_hash)),
-            Some((2, bank2_hash))
-        );
-        assert!((&heaviest_subtree_fork_choice)
-            .children(&(4, bank4_hash))
-            .unwrap()
-            .collect_vec()
-            .is_empty());
-    }
-
-    #[test]
-    fn test_simple_votes() {
-        // Init state
-        let mut vote_simulator = VoteSimulator::new(1);
-        let node_pubkey = vote_simulator.node_pubkeys[0];
-        let mut tower = Tower::default();
-
-        // Create the tree of banks
-        let forks = tr(0) / (tr(1) / (tr(2) / (tr(3) / (tr(4) / tr(5)))));
-
-        // Set the voting behavior
-        let mut cluster_votes = HashMap::new();
-        let votes = vec![1, 2, 3, 4, 5];
-        cluster_votes.insert(node_pubkey, votes.clone());
-        vote_simulator.fill_bank_forks(forks, &cluster_votes, true);
-
-        // Simulate the votes
-        for vote in votes {
-            assert!(vote_simulator
-                .simulate_vote(vote, &node_pubkey, &mut tower,)
-                .is_empty());
-        }
-
-        for i in 1..5 {
-            assert_eq!(tower.vote_state.votes[i - 1].slot() as usize, i);
-            assert_eq!(
-                tower.vote_state.votes[i - 1].confirmation_count() as usize,
-                6 - i
-            );
-        }
-    }
-
-    #[test]
-    fn test_switch_threshold_duplicate_rollback() {
-        run_test_switch_threshold_duplicate_rollback(false);
-    }
-
-    #[test]
-    #[should_panic]
-    fn test_switch_threshold_duplicate_rollback_panic() {
-        run_test_switch_threshold_duplicate_rollback(true);
-    }
-
-    fn setup_switch_test(num_accounts: usize) -> (Arc<Bank>, VoteSimulator, u64) {
-        // Init state
-        assert!(num_accounts > 1);
-        let mut vote_simulator = VoteSimulator::new(num_accounts);
-        let bank0 = vote_simulator.bank_forks.read().unwrap().get(0).unwrap();
-        let total_stake = bank0.total_epoch_stake();
-        assert_eq!(
-            total_stake,
-            vote_simulator.validator_keypairs.len() as u64 * 10_000
-        );
-
-        // Create the tree of banks
-        let forks = tr(0)
-            / (tr(1)
-                / (tr(2)
-                    // Minor fork 1
-                    / (tr(10) / (tr(11) / (tr(12) / (tr(13) / (tr(14))))))
-                    / (tr(43)
-                        / (tr(44)
-                            // Minor fork 2
-                            / (tr(45) / (tr(46) / (tr(47) / (tr(48) / (tr(49) / (tr(50)))))))
-                            / (tr(110)))
-                        / tr(112))));
-
-        // Fill the BankForks according to the above fork structure
-        vote_simulator.fill_bank_forks(forks, &HashMap::new(), true);
-        for (_, fork_progress) in vote_simulator.progress.iter_mut() {
-            fork_progress.fork_stats.computed = true;
-        }
-
-        (bank0, vote_simulator, total_stake)
-    }
-
-    fn run_test_switch_threshold_duplicate_rollback(should_panic: bool) {
-        let (bank0, mut vote_simulator, total_stake) = setup_switch_test(2);
-        let ancestors = vote_simulator.bank_forks.read().unwrap().ancestors();
-        let descendants = vote_simulator.bank_forks.read().unwrap().descendants();
-        let mut tower = Tower::default();
-
-        // Last vote is 47
-        tower.record_vote(
-            47,
-            vote_simulator
-                .bank_forks
-                .read()
-                .unwrap()
-                .get(47)
-                .unwrap()
-                .hash(),
-        );
-
-        // Trying to switch to an ancestor of last vote should only not panic
-        // if the current vote has a duplicate ancestor
-        let ancestor_of_voted_slot = 43;
-        let duplicate_ancestor1 = 44;
-        let duplicate_ancestor2 = 45;
-        vote_simulator
-            .heaviest_subtree_fork_choice
-            .mark_fork_invalid_candidate(&(
-                duplicate_ancestor1,
-                vote_simulator
-                    .bank_forks
-                    .read()
-                    .unwrap()
-                    .get(duplicate_ancestor1)
-                    .unwrap()
-                    .hash(),
-            ));
-        vote_simulator
-            .heaviest_subtree_fork_choice
-            .mark_fork_invalid_candidate(&(
-                duplicate_ancestor2,
-                vote_simulator
-                    .bank_forks
-                    .read()
-                    .unwrap()
-                    .get(duplicate_ancestor2)
-                    .unwrap()
-                    .hash(),
-            ));
-        assert_eq!(
-            tower.check_switch_threshold(
-                ancestor_of_voted_slot,
-                &ancestors,
-                &descendants,
-                &vote_simulator.progress,
-                total_stake,
-                bank0.epoch_vote_accounts(0).unwrap(),
-                &vote_simulator.latest_validator_votes_for_frozen_banks,
-                &vote_simulator.heaviest_subtree_fork_choice,
-            ),
-            SwitchForkDecision::FailedSwitchDuplicateRollback(duplicate_ancestor2)
-        );
-        let mut confirm_ancestors = vec![duplicate_ancestor1];
-        if should_panic {
-            // Adding the last duplicate ancestor will
-            // 1) Cause loop below to confirm last ancestor
-            // 2) Check switch threshold on a vote ancestor when there
-            // are no duplicates on that fork, which will cause a panic
-            confirm_ancestors.push(duplicate_ancestor2);
-        }
-        for (i, duplicate_ancestor) in confirm_ancestors.into_iter().enumerate() {
-            vote_simulator
-                .heaviest_subtree_fork_choice
-                .mark_fork_valid_candidate(&(
-                    duplicate_ancestor,
-                    vote_simulator
-                        .bank_forks
-                        .read()
-                        .unwrap()
-                        .get(duplicate_ancestor)
-                        .unwrap()
-                        .hash(),
-                ));
-            let res = tower.check_switch_threshold(
-                ancestor_of_voted_slot,
-                &ancestors,
-                &descendants,
-                &vote_simulator.progress,
-                total_stake,
-                bank0.epoch_vote_accounts(0).unwrap(),
-                &vote_simulator.latest_validator_votes_for_frozen_banks,
-                &vote_simulator.heaviest_subtree_fork_choice,
-            );
-            if i == 0 {
-                assert_eq!(
-                    res,
-                    SwitchForkDecision::FailedSwitchDuplicateRollback(duplicate_ancestor2)
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn test_switch_threshold() {
-        let (bank0, mut vote_simulator, total_stake) = setup_switch_test(2);
-        let ancestors = vote_simulator.bank_forks.read().unwrap().ancestors();
-        let mut descendants = vote_simulator.bank_forks.read().unwrap().descendants();
-        let mut tower = Tower::default();
-        let other_vote_account = vote_simulator.vote_pubkeys[1];
-
-        // Last vote is 47
-        tower.record_vote(47, Hash::default());
-
-        // Trying to switch to a descendant of last vote should always work
-        assert_eq!(
-            tower.check_switch_threshold(
-                48,
-                &ancestors,
-                &descendants,
-                &vote_simulator.progress,
-                total_stake,
-                bank0.epoch_vote_accounts(0).unwrap(),
-                &vote_simulator.latest_validator_votes_for_frozen_banks,
-                &vote_simulator.heaviest_subtree_fork_choice,
-            ),
-            SwitchForkDecision::SameFork
-        );
-
-        // Trying to switch to another fork at 110 should fail
-        assert_eq!(
-            tower.check_switch_threshold(
-                110,
-                &ancestors,
-                &descendants,
-                &vote_simulator.progress,
-                total_stake,
-                bank0.epoch_vote_accounts(0).unwrap(),
-                &vote_simulator.latest_validator_votes_for_frozen_banks,
-                &vote_simulator.heaviest_subtree_fork_choice,
-            ),
-            SwitchForkDecision::FailedSwitchThreshold(0, 20000)
-        );
-
-        // Adding another validator lockout on a descendant of last vote should
-        // not count toward the switch threshold
-        vote_simulator.simulate_lockout_interval(50, (49, 100), &other_vote_account);
-        assert_eq!(
-            tower.check_switch_threshold(
-                110,
-                &ancestors,
-                &descendants,
-                &vote_simulator.progress,
-                total_stake,
-                bank0.epoch_vote_accounts(0).unwrap(),
-                &vote_simulator.latest_validator_votes_for_frozen_banks,
-                &vote_simulator.heaviest_subtree_fork_choice,
-            ),
-            SwitchForkDecision::FailedSwitchThreshold(0, 20000)
-        );
-
-        // Adding another validator lockout on an ancestor of last vote should
-        // not count toward the switch threshold
-        vote_simulator.simulate_lockout_interval(50, (45, 100), &other_vote_account);
-        assert_eq!(
-            tower.check_switch_threshold(
-                110,
-                &ancestors,
-                &descendants,
-                &vote_simulator.progress,
-                total_stake,
-                bank0.epoch_vote_accounts(0).unwrap(),
-                &vote_simulator.latest_validator_votes_for_frozen_banks,
-                &vote_simulator.heaviest_subtree_fork_choice,
-            ),
-            SwitchForkDecision::FailedSwitchThreshold(0, 20000)
-        );
-
-        // Adding another validator lockout on a different fork, but the lockout
-        // doesn't cover the last vote, should not satisfy the switch threshold
-        vote_simulator.simulate_lockout_interval(14, (12, 46), &other_vote_account);
-        assert_eq!(
-            tower.check_switch_threshold(
-                110,
-                &ancestors,
-                &descendants,
-                &vote_simulator.progress,
-                total_stake,
-                bank0.epoch_vote_accounts(0).unwrap(),
-                &vote_simulator.latest_validator_votes_for_frozen_banks,
-                &vote_simulator.heaviest_subtree_fork_choice,
-            ),
-            SwitchForkDecision::FailedSwitchThreshold(0, 20000)
-        );
-
-        // Adding another validator lockout on a different fork, and the lockout
-        // covers the last vote would count towards the switch threshold,
-        // unless the bank is not the most recent frozen bank on the fork (14 is a
-        // frozen/computed bank > 13 on the same fork in this case)
-        vote_simulator.simulate_lockout_interval(13, (12, 47), &other_vote_account);
-        assert_eq!(
-            tower.check_switch_threshold(
-                110,
-                &ancestors,
-                &descendants,
-                &vote_simulator.progress,
-                total_stake,
-                bank0.epoch_vote_accounts(0).unwrap(),
-                &vote_simulator.latest_validator_votes_for_frozen_banks,
-                &vote_simulator.heaviest_subtree_fork_choice,
-            ),
-            SwitchForkDecision::FailedSwitchThreshold(0, 20000)
-        );
-
-        // Adding another validator lockout on a different fork, and the lockout
-        // covers the last vote, should satisfy the switch threshold
-        vote_simulator.simulate_lockout_interval(14, (12, 47), &other_vote_account);
-        assert_eq!(
-            tower.check_switch_threshold(
-                110,
-                &ancestors,
-                &descendants,
-                &vote_simulator.progress,
-                total_stake,
-                bank0.epoch_vote_accounts(0).unwrap(),
-                &vote_simulator.latest_validator_votes_for_frozen_banks,
-                &vote_simulator.heaviest_subtree_fork_choice,
-            ),
-            SwitchForkDecision::SwitchProof(Hash::default())
-        );
-
-        // Adding another unfrozen descendant of the tip of 14 should not remove
-        // slot 14 from consideration because it is still the most recent frozen
-        // bank on its fork
-        descendants.get_mut(&14).unwrap().insert(10000);
-        assert_eq!(
-            tower.check_switch_threshold(
-                110,
-                &ancestors,
-                &descendants,
-                &vote_simulator.progress,
-                total_stake,
-                bank0.epoch_vote_accounts(0).unwrap(),
-                &vote_simulator.latest_validator_votes_for_frozen_banks,
-                &vote_simulator.heaviest_subtree_fork_choice,
-            ),
-            SwitchForkDecision::SwitchProof(Hash::default())
-        );
-
-        // If we set a root, then any lockout intervals below the root shouldn't
-        // count toward the switch threshold. This means the other validator's
-        // vote lockout no longer counts
-        tower.vote_state.root_slot = Some(43);
-        // Refresh ancestors and descendants for new root.
-        let ancestors = vote_simulator.bank_forks.read().unwrap().ancestors();
-        let descendants = vote_simulator.bank_forks.read().unwrap().descendants();
-
-        assert_eq!(
-            tower.check_switch_threshold(
-                110,
-                &ancestors,
-                &descendants,
-                &vote_simulator.progress,
-                total_stake,
-                bank0.epoch_vote_accounts(0).unwrap(),
-                &vote_simulator.latest_validator_votes_for_frozen_banks,
-                &vote_simulator.heaviest_subtree_fork_choice,
-            ),
-            SwitchForkDecision::FailedSwitchThreshold(0, 20000)
-        );
-    }
-
-    #[test]
-    fn test_switch_threshold_use_gossip_votes() {
-        let num_validators = 2;
-        let (bank0, mut vote_simulator, total_stake) = setup_switch_test(2);
-        let ancestors = vote_simulator.bank_forks.read().unwrap().ancestors();
-        let descendants = vote_simulator.bank_forks.read().unwrap().descendants();
-        let mut tower = Tower::default();
-        let other_vote_account = vote_simulator.vote_pubkeys[1];
-
-        // Last vote is 47
-        tower.record_vote(47, Hash::default());
-
-        // Trying to switch to another fork at 110 should fail
-        assert_eq!(
-            tower.check_switch_threshold(
-                110,
-                &ancestors,
-                &descendants,
-                &vote_simulator.progress,
-                total_stake,
-                bank0.epoch_vote_accounts(0).unwrap(),
-                &vote_simulator.latest_validator_votes_for_frozen_banks,
-                &vote_simulator.heaviest_subtree_fork_choice,
-            ),
-            SwitchForkDecision::FailedSwitchThreshold(0, num_validators * 10000)
-        );
-
-        // Adding a vote on the descendant shouldn't count toward the switch threshold
-        vote_simulator.simulate_lockout_interval(50, (49, 100), &other_vote_account);
-        assert_eq!(
-            tower.check_switch_threshold(
-                110,
-                &ancestors,
-                &descendants,
-                &vote_simulator.progress,
-                total_stake,
-                bank0.epoch_vote_accounts(0).unwrap(),
-                &vote_simulator.latest_validator_votes_for_frozen_banks,
-                &vote_simulator.heaviest_subtree_fork_choice,
-            ),
-            SwitchForkDecision::FailedSwitchThreshold(0, 20000)
-        );
-
-        // Adding a later vote from gossip that isn't on the same fork should count toward the
-        // switch threshold
-        vote_simulator
-            .latest_validator_votes_for_frozen_banks
-            .check_add_vote(
-                other_vote_account,
-                112,
-                Some(
-                    vote_simulator
-                        .bank_forks
-                        .read()
-                        .unwrap()
-                        .get(112)
-                        .unwrap()
-                        .hash(),
-                ),
-                false,
-            );
-
-        assert_eq!(
-            tower.check_switch_threshold(
-                110,
-                &ancestors,
-                &descendants,
-                &vote_simulator.progress,
-                total_stake,
-                bank0.epoch_vote_accounts(0).unwrap(),
-                &vote_simulator.latest_validator_votes_for_frozen_banks,
-                &vote_simulator.heaviest_subtree_fork_choice,
-            ),
-            SwitchForkDecision::SwitchProof(Hash::default())
-        );
-
-        // If we now set a root that causes slot 112 to be purged from BankForks, then
-        // the switch proof will now fail since that validator's vote can no longer be
-        // included in the switching proof
-        vote_simulator.set_root(44);
-        let ancestors = vote_simulator.bank_forks.read().unwrap().ancestors();
-        let descendants = vote_simulator.bank_forks.read().unwrap().descendants();
-        assert_eq!(
-            tower.check_switch_threshold(
-                110,
-                &ancestors,
-                &descendants,
-                &vote_simulator.progress,
-                total_stake,
-                bank0.epoch_vote_accounts(0).unwrap(),
-                &vote_simulator.latest_validator_votes_for_frozen_banks,
-                &vote_simulator.heaviest_subtree_fork_choice,
-            ),
-            SwitchForkDecision::FailedSwitchThreshold(0, 20000)
-        );
-    }
-
-    #[test]
-    fn test_switch_threshold_votes() {
-        // Init state
-        let mut vote_simulator = VoteSimulator::new(4);
-        let node_pubkey = vote_simulator.node_pubkeys[0];
-        let mut tower = Tower::default();
-        let forks = tr(0)
-            / (tr(1)
-                / (tr(2)
-                    // Minor fork 1
-                    / (tr(10) / (tr(11) / (tr(12) / (tr(13) / (tr(14))))))
-                    / (tr(43)
-                        / (tr(44)
-                            // Minor fork 2
-                            / (tr(45) / (tr(46))))
-                        / (tr(110)))));
-
-        // Have two validators, each representing 20% of the stake vote on
-        // minor fork 2 at slots 46 + 47
-        let mut cluster_votes: HashMap<Pubkey, Vec<Slot>> = HashMap::new();
-        cluster_votes.insert(vote_simulator.node_pubkeys[1], vec![46]);
-        cluster_votes.insert(vote_simulator.node_pubkeys[2], vec![47]);
-        vote_simulator.fill_bank_forks(forks, &cluster_votes, true);
-
-        // Vote on the first minor fork at slot 14, should succeed
-        assert!(vote_simulator
-            .simulate_vote(14, &node_pubkey, &mut tower,)
-            .is_empty());
-
-        // The other two validators voted at slots 46, 47, which
-        // will only both show up in slot 48, at which point
-        // 2/5 > SWITCH_FORK_THRESHOLD of the stake has voted
-        // on another fork, so switching should succeed
-        let votes_to_simulate = (46..=48).collect();
-        let results = vote_simulator.create_and_vote_new_branch(
-            45,
-            48,
-            &cluster_votes,
-            &votes_to_simulate,
-            &node_pubkey,
-            &mut tower,
-        );
-        assert_eq!(
-            *results.get(&46).unwrap(),
-            vec![HeaviestForkFailures::FailedSwitchThreshold(46, 0, 40000)]
-        );
-        assert_eq!(
-            *results.get(&47).unwrap(),
-            vec![HeaviestForkFailures::FailedSwitchThreshold(
-                47, 10000, 40000
-            )]
-        );
-        assert!(results.get(&48).unwrap().is_empty());
-    }
-
-    #[test]
-    fn test_double_partition() {
-        // Init state
-        let mut vote_simulator = VoteSimulator::new(2);
-        let node_pubkey = vote_simulator.node_pubkeys[0];
-        let vote_pubkey = vote_simulator.vote_pubkeys[0];
-        let mut tower = Tower::default();
-
-        let num_slots_to_try = 200;
-        // Create the tree of banks
-        let forks = tr(0)
-            / (tr(1)
-                / (tr(2)
-                    / (tr(3)
-                        / (tr(4)
-                            / (tr(5)
-                                / (tr(6)
-                                    / (tr(7)
-                                        / (tr(8)
-                                            / (tr(9)
-                                                // Minor fork 1
-                                                / (tr(10) / (tr(11) / (tr(12) / (tr(13) / (tr(14))))))
-                                                / (tr(43)
-                                                    / (tr(44)
-                                                        // Minor fork 2
-                                                        / (tr(45) / (tr(46) / (tr(47) / (tr(48) / (tr(49) / (tr(50)))))))
-                                                        / (tr(110) / (tr(110 + 2 * num_slots_to_try))))))))))))));
-
-        // Set the successful voting behavior
-        let mut cluster_votes = HashMap::new();
-        let mut my_votes: Vec<Slot> = vec![];
-        let next_unlocked_slot = 110;
-        // Vote on the first minor fork
-        my_votes.extend(1..=14);
-        // Come back to the main fork
-        my_votes.extend(43..=44);
-        // Vote on the second minor fork
-        my_votes.extend(45..=50);
-        // Vote to come back to main fork
-        my_votes.push(next_unlocked_slot);
-        cluster_votes.insert(node_pubkey, my_votes.clone());
-        // Make the other validator vote fork to pass the threshold checks
-        let other_votes = my_votes.clone();
-        cluster_votes.insert(vote_simulator.node_pubkeys[1], other_votes);
-        vote_simulator.fill_bank_forks(forks, &cluster_votes, true);
-
-        // Simulate the votes.
-        for vote in &my_votes {
-            // All these votes should be ok
-            assert!(vote_simulator
-                .simulate_vote(*vote, &node_pubkey, &mut tower,)
-                .is_empty());
-        }
-
-        info!("local tower: {:#?}", tower.vote_state.votes);
-        let observed = vote_simulator
-            .bank_forks
-            .read()
-            .unwrap()
-            .get(next_unlocked_slot)
-            .unwrap()
-            .get_vote_account(&vote_pubkey)
-            .unwrap();
-        let state = observed.vote_state();
-        info!("observed tower: {:#?}", state.as_ref().unwrap().votes);
-
-        let num_slots_to_try = 200;
-        cluster_votes
-            .get_mut(&vote_simulator.node_pubkeys[1])
-            .unwrap()
-            .extend(next_unlocked_slot + 1..next_unlocked_slot + num_slots_to_try);
-        assert!(vote_simulator.can_progress_on_fork(
-            &node_pubkey,
-            &mut tower,
-            next_unlocked_slot,
-            num_slots_to_try,
-            &mut cluster_votes,
-        ));
-    }
-
-    #[test]
-    fn test_switch_threshold_across_tower_reload() {
-        solana_logger::setup();
-        // Init state
-        let mut vote_simulator = VoteSimulator::new(2);
-        let other_vote_account = vote_simulator.vote_pubkeys[1];
-        let bank0 = vote_simulator.bank_forks.read().unwrap().get(0).unwrap();
-        let total_stake = bank0.total_epoch_stake();
-        assert_eq!(
-            total_stake,
-            vote_simulator.validator_keypairs.len() as u64 * 10_000
-        );
-
-        // Create the tree of banks
-        let forks = tr(0)
-            / (tr(1)
-                / (tr(2)
-                    / tr(10)
-                    / (tr(43)
-                        / (tr(44)
-                            // Minor fork 2
-                            / (tr(45) / (tr(46) / (tr(47) / (tr(48) / (tr(49) / (tr(50)))))))
-                            / (tr(110) / tr(111))))));
-
-        // Fill the BankForks according to the above fork structure
-        vote_simulator.fill_bank_forks(forks, &HashMap::new(), true);
-        for (_, fork_progress) in vote_simulator.progress.iter_mut() {
-            fork_progress.fork_stats.computed = true;
-        }
-
-        let ancestors = vote_simulator.bank_forks.read().unwrap().ancestors();
-        let descendants = vote_simulator.bank_forks.read().unwrap().descendants();
-        let mut tower = Tower::default();
-
-        tower.record_vote(43, Hash::default());
-        tower.record_vote(44, Hash::default());
-        tower.record_vote(45, Hash::default());
-        tower.record_vote(46, Hash::default());
-        tower.record_vote(47, Hash::default());
-        tower.record_vote(48, Hash::default());
-        tower.record_vote(49, Hash::default());
-
-        // Trying to switch to a descendant of last vote should always work
-        assert_eq!(
-            tower.check_switch_threshold(
-                50,
-                &ancestors,
-                &descendants,
-                &vote_simulator.progress,
-                total_stake,
-                bank0.epoch_vote_accounts(0).unwrap(),
-                &vote_simulator.latest_validator_votes_for_frozen_banks,
-                &vote_simulator.heaviest_subtree_fork_choice,
-            ),
-            SwitchForkDecision::SameFork
-        );
-
-        // Trying to switch to another fork at 110 should fail
-        assert_eq!(
-            tower.check_switch_threshold(
-                110,
-                &ancestors,
-                &descendants,
-                &vote_simulator.progress,
-                total_stake,
-                bank0.epoch_vote_accounts(0).unwrap(),
-                &vote_simulator.latest_validator_votes_for_frozen_banks,
-                &vote_simulator.heaviest_subtree_fork_choice,
-            ),
-            SwitchForkDecision::FailedSwitchThreshold(0, 20000)
-        );
-
-        vote_simulator.simulate_lockout_interval(111, (10, 49), &other_vote_account);
-
-        assert_eq!(
-            tower.check_switch_threshold(
-                110,
-                &ancestors,
-                &descendants,
-                &vote_simulator.progress,
-                total_stake,
-                bank0.epoch_vote_accounts(0).unwrap(),
-                &vote_simulator.latest_validator_votes_for_frozen_banks,
-                &vote_simulator.heaviest_subtree_fork_choice,
-            ),
-            SwitchForkDecision::SwitchProof(Hash::default())
-        );
-
-        assert_eq!(tower.voted_slots(), vec![43, 44, 45, 46, 47, 48, 49]);
-        {
-            let mut tower = tower.clone();
-            tower.record_vote(110, Hash::default());
-            tower.record_vote(111, Hash::default());
-            assert_eq!(tower.voted_slots(), vec![43, 110, 111]);
-            assert_eq!(tower.vote_state.root_slot, Some(0));
-        }
-
-        // Prepare simulated validator restart!
-        let mut vote_simulator = VoteSimulator::new(2);
-        let other_vote_account = vote_simulator.vote_pubkeys[1];
-        let bank0 = vote_simulator.bank_forks.read().unwrap().get(0).unwrap();
-        let total_stake = bank0.total_epoch_stake();
-        let forks = tr(0)
-            / (tr(1)
-                / (tr(2)
-                    / tr(10)
-                    / (tr(43)
-                        / (tr(44)
-                            // Minor fork 2
-                            / (tr(45) / (tr(46) / (tr(47) / (tr(48) / (tr(49) / (tr(50)))))))
-                            / (tr(110) / tr(111))))));
-        let replayed_root_slot = 44;
-
-        // Fill the BankForks according to the above fork structure
-        vote_simulator.fill_bank_forks(forks, &HashMap::new(), true);
-        for (_, fork_progress) in vote_simulator.progress.iter_mut() {
-            fork_progress.fork_stats.computed = true;
-        }
-
-        // prepend tower restart!
-        let mut slot_history = SlotHistory::default();
-        vote_simulator.set_root(replayed_root_slot);
-        let ancestors = vote_simulator.bank_forks.read().unwrap().ancestors();
-        let descendants = vote_simulator.bank_forks.read().unwrap().descendants();
-        for slot in &[0, 1, 2, 43, replayed_root_slot] {
-            slot_history.add(*slot);
-        }
-        let mut tower = tower
-            .adjust_lockouts_after_replay(replayed_root_slot, &slot_history)
-            .unwrap();
-
-        assert_eq!(tower.voted_slots(), vec![45, 46, 47, 48, 49]);
-
-        // Trying to switch to another fork at 110 should fail
-        assert_eq!(
-            tower.check_switch_threshold(
-                110,
-                &ancestors,
-                &descendants,
-                &vote_simulator.progress,
-                total_stake,
-                bank0.epoch_vote_accounts(0).unwrap(),
-                &vote_simulator.latest_validator_votes_for_frozen_banks,
-                &vote_simulator.heaviest_subtree_fork_choice,
-            ),
-            SwitchForkDecision::FailedSwitchThreshold(0, 20000)
-        );
-
-        // Add lockout_interval which should be excluded
-        vote_simulator.simulate_lockout_interval(111, (45, 50), &other_vote_account);
-        assert_eq!(
-            tower.check_switch_threshold(
-                110,
-                &ancestors,
-                &descendants,
-                &vote_simulator.progress,
-                total_stake,
-                bank0.epoch_vote_accounts(0).unwrap(),
-                &vote_simulator.latest_validator_votes_for_frozen_banks,
-                &vote_simulator.heaviest_subtree_fork_choice,
-            ),
-            SwitchForkDecision::FailedSwitchThreshold(0, 20000)
-        );
-
-        // Add lockout_interval which should not be excluded
-        vote_simulator.simulate_lockout_interval(111, (110, 200), &other_vote_account);
-        assert_eq!(
-            tower.check_switch_threshold(
-                110,
-                &ancestors,
-                &descendants,
-                &vote_simulator.progress,
-                total_stake,
-                bank0.epoch_vote_accounts(0).unwrap(),
-                &vote_simulator.latest_validator_votes_for_frozen_banks,
-                &vote_simulator.heaviest_subtree_fork_choice,
-            ),
-            SwitchForkDecision::SwitchProof(Hash::default())
-        );
-
-        tower.record_vote(110, Hash::default());
-        tower.record_vote(111, Hash::default());
-        assert_eq!(tower.voted_slots(), vec![110, 111]);
-        assert_eq!(tower.vote_state.root_slot, Some(replayed_root_slot));
-    }
 }
