+++ local-cluster/Cargo.toml
@@ -18,6 +18,7 @@ rayon = { workspace = true }
 solana-accounts-db = { workspace = true }
 solana-client = { workspace = true }
 solana-config-program = { workspace = true }
+solana-consensus = { workspace = true }
 solana-core = { workspace = true }
 solana-entry = { workspace = true }
 solana-gossip = { workspace = true }
