+++ local-cluster/src/local_cluster.rs
@@ -8,10 +8,8 @@ use {
     log::*,
     solana_accounts_db::accounts_db::create_accounts_run_and_snapshot_dirs,
     solana_client::{connection_cache::ConnectionCache, thin_client::ThinClient},
-    solana_core::{
-        consensus::tower_storage::FileTowerStorage,
-        validator::{Validator, ValidatorConfig, ValidatorStartProgress},
-    },
+    solana_consensus::tower_storage::FileTowerStorage,
+    solana_core::validator::{Validator, ValidatorConfig, ValidatorStartProgress},
     solana_gossip::{
         cluster_info::Node,
         contact_info::{ContactInfo, LegacyContactInfo, Protocol},
