+++ validator/src/main.rs
@@ -19,9 +19,9 @@ use {
         partitioned_rewards::TestPartitionedEpochRewards,
     },
     solana_clap_utils::input_parsers::{keypair_of, keypairs_of, pubkey_of, value_of},
+    solana_consensus::tower_storage,
     solana_core::{
         banking_trace::DISABLED_BAKING_TRACE_DIR,
-        consensus::tower_storage,
         system_monitor_service::SystemMonitorService,
         tpu::DEFAULT_TPU_COALESCE,
         validator::{
